+++ packages/kit-headless/src/components/autocomplete/autocomplete-context-id.ts
@@ -1,0 +1,7 @@
+import { createContextId } from '@builder.io/qwik';
+import { AutocompleteContext } from './autocomplete-context.type';
+
+const AutocompleteContextId =
+  createContextId<AutocompleteContext>('autocomplete-root');
+
+export default AutocompleteContextId;
