+++ packages/kit-headless/src/components/autocomplete/autocomplete.spec.tsx
@@ -7,7 +7,7 @@ import {
   AutocompleteButton,
   AutocompleteListbox,
   AutocompleteOption,
-} from './autocomplete';
+} from './index';
 
 const fruits = [
   'Apple',
