+++ packages/kit-headless/src/components/autocomplete/index.ts
@@ -1,0 +1,10 @@
+export * from './autocomplete-context-id';
+export * from './autocomplete-context.type';
+export * from './autocomplete-root';
+export * from './autocomplete-label';
+export * from './autocomplete-trigger';
+export * from './autocomplete-input';
+export * from './autocomplete-button';
+export * from './autocomplete-listbox';
+// export * from './autocomplete-group';
+export * from './autocomplete-option';
